requirements.txt
alpaca-trade-api==3.1.1
pandas>=1.5.0
numpy>=1.21.0
python-dotenv>=0.19.0
